pyproject.toml
[project]
name = "umspages"
version = "0.2.0"
description = "Page Objects used for end to end testing in the SouvAP Dev team"
maintainers = [
    {name = "Univention GmbH", email = "packages@univention.de"},
]
license = {text = "GNU Affero General Public License v3"}
readme = "README.md"
requires-python = ">=3.8,<4"

[project.optional-dependencies]
tests = [ "pytest-playwright==0.4.2", "requests==2.31.0", "url-normalize==1.4.3", "pytest-html==4.0.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/umspages", "src/pom"]

[tool.pytest.ini_options]
pythonpath = [ "src" ]
testpaths = [ "tests" ]
markers = [
    "saml: all tests that enforce SAML login. Other tests may also use SAML depending on the environment",
    "gaia: marks tests that are used for testing Gaia cluster",
    "devenv: marks tests that are used for testing Dev-env cluster",
]
